pyproject.toml
[tool.poetry]
name = "sport-data-lambda-trigger"
version = "0.1.0"
description = ""
authors = ["Fakile Temitope Oluwaseyi <fakiletemitope@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
boto3 = "^1.35.94"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
